FIP64: Flow translator handles to VPP action
This allows the flow translator to handle ToVppAction virtual
actions, which will be translated to the packet being tunneled
to the gateway host using the action's VNI.

Ref: MI-1730

+++ midolman/src/main/scala/org/midonet/midolman/simulation/Simulator.scala
@@ -40,7 +40,7 @@ object Simulator {
       * processing in a VPP pipeline. The action includes the identifier of the
       * VPP host and the VNI of the VXLAN tunnel.
       */
-    case class ToVppAction(hostId: UUID, vni: Int)
+    case class ToVppAction(hostId: UUID, vni: Long)
         extends ForwardAction with VirtualFlowAction
 
     /**
